The intial commit for the Miami-Properties Project, containing the Property Boundary View Dataset from Miami-Dade County's Open Data Hub.
The python script calls on the Miami-Properties API, intializes a CSV in your local directory with the data received from the API, and reads the CSV into a Pandas Dataframe.

The script performs some Data Processing on the large dataset, and downloads a csv file (single_family.csv) with Residential-Single Family Properties which serve as the basis for our intial data analysis.